Remove caching of map instance
Even though this sounds like it should have performance benefits, it
causes issues when navigating with the Back button, and it also prevents
us from having multiple map instances.

This commit removes the caching (the map gets reinitialized every time
the component gets instantiated).

Also, it fixes the hard coded div ID (inspired on
https://gist.github.com/ZogStriP/5684983) and instead uses the container
for the component itself as the canvas.